+++ CHANGELOG.md
@@ -18,8 +18,8 @@ noise ([#58](https://github.com/ctrltz/meegsim/pull/58))
 - Allow specifying standard deviation via a SourceEstimate object ([#67](https://github.com/ctrltz/meegsim/pull/67))
 - A method for setting phase-phase coupling by adding noise to the shifted copy of input waveform ([#71](https://github.com/ctrltz/meegsim/pull/71))
 - Function to convert the sources to mne.Label ([#73](https://github.com/ctrltz/meegsim/pull/73))
-- Quick list-like access to the simulated sources ([#82](https://github.com/ctrltz/meegsim/pull/82))
-- Control over the amplitude envelope of the coupled waveform ([#87](https://github.com/ctrltz/meegsim/pull/87))
+- Quick dict-like access to the simulated sources ([#82](https://github.com/ctrltz/meegsim/pull/82))
+- Partial control over the amplitude envelope of the coupled waveform: same as input or randomly generated ([#87](https://github.com/ctrltz/meegsim/pull/87))
 
 ### Changed